Fix the compilation with old 2.6 kernels
    7216afbf27f249a4f30d02a4cfa3223e4a0672ee
pdplus - fix a compile warning
    692c86062a966257b27c8ba2fd1422fe617b776e

+++ alsa/pci/pdplus/pdplus.c
@@ -6162,7 +6162,7 @@ static void __devexit pdplus_remove(pci_dev_t *pci)
         struct snd_card *card;
         ENTER;
 
-        Vprintk ("MMIO 0= 0x%lx\n", pci_resource_start (pci, 0));
+        Vprintk ("MMIO 0= 0x%lx\n", (unsigned long)pci_resource_start (pci, 0));
         card = pci_get_drvdata (pci);
 
         if (card == 0)
